feat: OptionDetailDialog UI구현
