+++ Android/app/src/main/res/values/dimens.xml
@@ -1,5 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <dimen name="pageMargin">70dp</dimen>
-    <dimen name="pagerWidth">312dp</dimen>
+    <dimen name="page_margin">70dp</dimen>
+    <dimen name="pager_width">312dp</dimen>
+
+    <dimen name="option_detail_dialog_width">380dp</dimen>
+    <dimen name="option_detail_dialog_height">650dp</dimen>
+
 </resources>
