+++ Android/app/src/main/res/values/styles.xml
@@ -31,6 +31,11 @@
         <item name="android:textSize">18sp</item>
     </style>
 
+    <style name="TabOptionDetailTextStyle" parent="TabTextStyle">
+        <item name="android:textSize">14sp</item>
+        <item name="fontFamily">@font/hyndaisans_text</item>
+    </style>
+
     <style name="TabItemStyle" parent="@android:style/TextAppearance.Widget.TabWidget">
         <item name="android:fontFamily">@font/hyndaisans_head_bold</item>
         <item name="android:textSize">14sp</item>
